Document pipeline service extraction scope
